==> 1_build_and_preprocess.ipynb <==
## Preprocessing and building dataset

We perform the described preprocessing, then we build a DataFrame with preprocessed tweets.

### Configuration presets

import pandas as pd
from sklearn.model_selection import train_test_split

PREFIX = './datasets'
LABEL_MAPPER = {
    'antichina': 0,
    'antivacina': 1,
    'provacina': 2,
}

LOWERCASE = True
RANDOM_SEED = 42
OUTPUT_DATASET_SIZE_BY_CLASS = 1000
OUTPUT_FILE_NAME = '3000'

### Preprocessing function

import re
import unidecode
import emoji

hashtags = [
    '#EuVouTomarVacina',
    '#VacinaBrasil',
    '#VacinaEAmorAoPróximo',
    '#VacinaEAmorAoProximo', # without accent
    '#VacinaJá',
    '#VacinaJa', # without accent
    '#VacinaNoBrasil',
    '#VacinaParaTodos',
    '#VacinasPelaVida',
    '#VacinaUrgenteParaTodos',
    '#VemVacina',
    '#EuNãoVouTomarVacina',
    '#EuNaoVouTomarVacina', # withouth accent
    '#NãoVouTomarVacina',
    '#NaoVouTomarVacina', # without accent
    '#VacinaNão',
    '#VacinaNao', # without accent
    '#VacinaObrigatóriaNão',
    '#VacinaObrigatóriaNao', # without accent
    '#VacinaObrigatoriaNão', # without accent
    '#VacinaObrigatoriaNao', # without accent
    '#VachinaNão',
    '#VachinaNao', # without accent
    '#VachinaNãoPresidente',
    '#VachinaNaoPresidente', # without accent
    '#VachinaObrigatóriaNão',
    '#VachinaObrigatóriaNao', # without accent
    '#VachinaObrigatoriaNão', # without accent
    '#VachinaObrigatoriaNao', # without accent
    '#VacinaChinesaNão',
    '#VacinaChinesaNao', # without accent
]


def removeNonLatinCharacters(tweet):
    return " ".join(
        regex.sub(r'[^\p{Latin}]', u'', tweet).split()
    )


def removeNumbers(tweet):
    return " ".join(
        re.sub(r'[0-9]', '', tweet).split()
    )


def removeEmojis(tweet):
    return " ".join(
        emoji.get_emoji_regexp().sub(r'', tweet).split()
    )


def removeURLs(tweet):
    return " ".join(
        re.sub(r"http\S+|youtu.be\S+|\S+.com.br\S+|bit.ly\S+|\S+.com/\S+|\S+.co/\S+|\S+.org\S+|\S+.br/\S+|\S+.es/S+", "", tweet).split()
    )


def removeHashtags(tweet):
    for hashtag in hashtags:
        # Here, it could be two variations:
        # the hashtag symbol (#) + the HT
        # the HT only
        ht = re.compile(
            re.escape(
                hashtag
            ),
            re.IGNORECASE
        )
        tweet = ht.sub('', tweet)
        
        ht = re.compile(
            re.escape(
                hashtag[1:]
            ),
            re.IGNORECASE
        )
        tweet = ht.sub('', tweet)
    
    # Finally, replace all empty hashtags symbols (#)
    return " ".join(
        re.sub(r'# ', '', tweet).split()
    )


def removeMentions(tweet):
    return " ".join(
        re.sub(r"@\S+", "", tweet).split()
    )


def isValidTweet(tweet):
    without_mentions = " ".join(re.sub(r"@", "", tweet).split())
    if (
        len(without_mentions.split()) < 3 or
        not without_mentions or
        re.search("^\s*$", without_mentions)
    ):
        return False
    return True


def preprocessTweet(tweet, antivacina=True):
    preprocessed = removeNumbers(
        removeEmojis(
            removeMentions(
                removeURLs(
                    removeHashtags(tweet)
                )
            )
        )
    )
    if isValidTweet(preprocessed):
        if LOWERCASE:
            return preprocessed.lower()
        else:
            return preprocessed
    return None

# This is just for testing...
antichina = pd.read_csv(f'{PREFIX}/raw_csv/antichina.csv', names=['tweet'])
print(antichina['tweet'].values[1435])
print(preprocessTweet(antichina['tweet'].values[1435]))
print('')

antivax = pd.read_csv(f'{PREFIX}/raw_csv/antivacina.csv', names=['tweet'])
print(antivax['tweet'].values[37])
print(preprocessTweet(antivax['tweet'].values[37]))
print('')

provax = pd.read_csv(f'{PREFIX}/raw_csv/provacina.csv', names=['tweet'])
print(provax['tweet'].values[45])
print(preprocessTweet(provax['tweet'].values[45]))

### Dataset building

# Load and build dataset
print('Building dataset...\n')

labels = []
tweets = []

for label in ['antichina', 'antivacina', 'provacina']:
    print(f'Preprocessing {label}...')
    
    group_df = pd.read_csv(f'{PREFIX}/raw_csv/{label}.csv', names=['tweet'])
    
    for tweet in group_df['tweet'].values:
        try:
            preprocessed_tweet = preprocessTweet(tweet)
        except:
            print(f'Failed with {tweet}')
            continue

        if preprocessed_tweet:
            tweets.append(preprocessed_tweet)
            labels.append(LABEL_MAPPER[label])
    
    print('Finished.\n')
    
dataset = pd.DataFrame({
    'tweet': tweets,
    'label': labels,
})
dataset.to_csv(f'{PREFIX}/complete_dataset.csv', sep=';', columns=['tweet', 'label'], index=False)
print(dataset)

print('Complete dataset metrics...')
print('# Antichina:', len(dataset[dataset['label'] == 0]))
print('# Antivacina:', len(dataset[dataset['label'] == 1]))
print('# Provacina:', len(dataset[dataset['label'] == 2]))

# Now build a with the same size of elements of each class
sample = pd.concat([
    dataset[dataset['label'] == 0].sample(random_state=RANDOM_SEED, n=OUTPUT_DATASET_SIZE_BY_CLASS).reset_index(drop=True),
    dataset[dataset['label'] == 1].sample(random_state=RANDOM_SEED, n=OUTPUT_DATASET_SIZE_BY_CLASS).reset_index(drop=True),
    dataset[dataset['label'] == 2].sample(random_state=RANDOM_SEED, n=OUTPUT_DATASET_SIZE_BY_CLASS).reset_index(drop=True),
]).reset_index(drop=True)
sample.to_csv(f'{PREFIX}/{OUTPUT_FILE_NAME}.csv', sep=';', columns=['tweet', 'label'], index=False)
sample

print('Sample dataset metrics...')
print('# Antichina:', len(sample[sample['label'] == 0]))
print('# Antivacina:', len(sample[sample['label'] == 1]))
print('# Provacina:', len(sample[sample['label'] == 2]))
